Move AVM interfaces to their own file.
